+++ src/components/Game.jsx
@@ -3,9 +3,8 @@ import React, { Suspense } from "react";
 // Import components
 import Car from "./Car";
 import Obstacles from "./Obstacles";
-//import Sun from "./Sun";
 import Background from "./Background/index";
-import Environment from "./Environment";
+import Environment from "./Environment/index";
 
 // Import hooks
 import usePlayerPosition from "../hooks/usePlayerPosition";
